Updated Player Guides DR Mod on Report - Issue #59
Revised language about Modifiers to re-specify that the +X modifier gets added to the DR when making a check.

This is based on issue #59 and the general suggestion from playtest feedback to make the Modifiers and Target Card system feel more natural early on in playing the game.

Also minor fix to Poisoned to drain Stamina instead of HP (Issue #73).

Added Fatigue as an explicit Status Condition. (partially addressing Issue #69).

Updated 'Primary Suit' to 'Burn Mechanics' (addressing Issue #78)