Second Pass Update on Bestiary v1
Removed Dealers, added notes on default Range and to hit.

Added additional Actions to enemies and streamlined naming conventions to make Setting agnostic. This approach will give us a template to work off of for creatures of various power and difficulty without needing to bloat an initial Bestiary.

Will tackle Dealers in a separate pass.