Added Powers for Necromancer Boss
Added Ally Swap, similar to Phase Shift but also grants the swapped target one free attack.

Added Undying (Passive) and Undying (Vulnerability) - not sure if we have 'Boss-Only' tag in Vulnys but we can move that there if it makes more sense. Wanted it to be GM free to decide the nature of the Undying and the corresponding Vulny but open to modifying language as needed.