Update 02_GMGuide.md
Working to address #119 - rather than solve for every individual house rule instead add a blanket statement up front in the GM guide encouraging GM's to modify the rules as they see fit to meet their and their players' individual needs and goals.

Open to feedback on the placement or wording.