Updated Player Guide Brief for ease of onboarding
Updated the PGF to make terminology more friendly and accessible to new players. Removed instances of "TC" and "DR" to have the full terms read each time, and added more context to certain mechanics descriptions.

Operating under the assumption this single guide can be read and then have the players engage in a full session, the goal is to ensure it makes sense and can be understood by players with varying degrees of experience in the TTRPG genre.

As I went through I also made sure to correctly format, which resulted in some larger diffs based on modifying the text and formatting it to fit on new lines.